improves the code
